Fix // newline handling in freetype font (#2045)
* Fix // newline handling in freetype font

Fixes #2043

* Fix white space at start of lines in freetype

Skipping whitespace was checking the current character (past the line
length) not the character after the last breakpoint which is the real
start of the new line.

Fixes #2046

* Fix off by one error on space removal

We want to remove when the next character is a space not when the
current character is a space.

* Fix end of string when skipping white space

* Split out string functions and unit test
